+++ src/lib/encodings/base64.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2020 Angus.Fenying <fenying@litert.org>
+ * Copyright 2024 Angus.Fenying <fenying@litert.org>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,38 @@
  * limitations under the License.
  */
 
+/**
+ * Encode a string into a BASE64 encoding string.
+ *
+ * @param data  The string to be encoded.
+ * @returns     The BASE64 encoded string.
+ */
 export function stringToBase64(data: string): string {
 
     return Buffer.from(data).toString('base64');
 }
 
+/**
+ * Decode a BASE64 encoding string into a string.
+ */
 export function stringFromBase64(data: string): string {
 
     return Buffer.from(data, 'base64').toString();
 }
 
+/**
+ * Encode a buffer to a BASE64 encoding string.
+ */
 export function bufferToBase64(data: Buffer): string {
 
     return data.toString('base64');
 }
 
+/**
+ * Decode a BASE64 encoding string into a buffer.
+ *
+ * @param data  The BASE64 encoded string to be decoded.
+ */
 export function bufferFromBase64(data: string): Buffer {
 
     return Buffer.from(data, 'base64');