+++ src/lib/encodings/base64url.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2020 Angus.Fenying <fenying@litert.org>
+ * Copyright 2024 Angus.Fenying <fenying@litert.org>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-import * as Base64 from './base64';
-
 /**
  * Convert a string from BASE64 to BASE64-URL-Safe encoding.
  *
- * @param {string} base64 The BASE64 encoding string to be converted.
+ * @param {string} base64str The BASE64 encoding string to be converted.
  *
  * @returns {string} returns the BASE64-URL-Safe encoding result of input.
  */
-export function base64UrlEncode(base64: string): string {
+export function base64UrlEncode(base64str: string): string {
 
-    return base64
+    return base64str
         .replace(/=/g, '')
         .replace(/\+/g, '-')
         .replace(/\//g, '_');
@@ -34,36 +32,61 @@ export function base64UrlEncode(base64: string): string {
 /**
  * Convert a string from BASE64-URL-Safe encoding to BASE64.
  *
- * @param {string} base64 The BASE64-URL-Safe encoding string to be converted.
+ * @param {string} base64urlStr The BASE64-URL-Safe encoding string to be converted.
  *
  * @returns {string} returns the BASE64 encoding result of input.
  */
-export function base64UrlDecode(urlbase64: string): string {
+export function base64UrlDecode(base64urlStr: string): string {
 
-    const r = urlbase64.length % 4;
+    const r = base64urlStr.length % 4;
 
-    return urlbase64
+    return base64urlStr
         .replace(/-/g, '+')
         .replace(/_/g, '/')
-        .padEnd(urlbase64.length + (r ? (4 - r) : 0), '=');
+        .padEnd(base64urlStr.length + (r ? (4 - r) : 0), '=');
 }
 
+/**
+ * Encode a string to BASE64-URL-Safe encoding string.
+ *
+ * @param data  The BASE64 encoded string to be converted.
+ *
+ * @returns     The BASE64-URL-Safe encoded string.
+ */
 export function stringToBase64Url(data: string): string {
 
-    return base64UrlEncode(Base64.stringToBase64(data));
+    return Buffer.from(data).toString('base64url');
 }
 
+/**
+ * Decode BASE64-URL-Safe encoding string to string.
+ *
+ * @param data  The BASE64-URL-Safe encoded string to be converted.
+ * @returns     The BASE64 encoded string.
+ */
 export function stringFromBase64Url(data: string): string {
 
-    return Base64.stringFromBase64(base64UrlDecode(data));
+    return Buffer.from(data, 'base64url').toString();
 }
 
+/**
+ * Encode a buffer to BASE64-URL-Safe encoding string.
+ *
+ * @param data  The BASE64 encoded buffer to be converted.
+ * @returns     The BASE64-URL-Safe encoded string.
+ */
 export function bufferToBase64Url(data: Buffer): string {
 
-    return base64UrlEncode(Base64.bufferToBase64(data));
+    return data.toString('base64url');
 }
 
+/**
+ * Decode BASE64-URL-Safe encoding string to Buffer.
+ *
+ * @param data  The BASE64-URL-Safe encoded buffer to be converted.
+ * @returns     The BASE64 encoded buffer.
+ */
 export function bufferFromBase64Url(data: string): Buffer {
 
-    return Base64.bufferFromBase64(base64UrlDecode(data));
+    return Buffer.from(data, 'base64url');
 }